[COMPLETE] Webpack code split, README: added Ion's name
Splitted JS code with new webpack config, added Ion's name to README.md, more info here https://vimeo.com/504823763
