+++ package.json
@@ -1,33 +1,47 @@
 {
     "name": "team-four-webapp-frontend",
     "version": "1.0.0",
-    "description": "",
+    "description": "Responsive Web App PhotoMatch Frontend",
     "main": "index.js",
     "scripts": {
         "dev": "webpack --mode development",
         "start": "webpack serve --open Chrome.exe",
         "build": "webpack --mode production"
     },
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "lite-server": "^2.6.1"
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/manuelalonge/team-four-webapp-frontend.git"
     },
+    "keywords": [
+        "Responsive",
+        "Frontend",
+        "Webapp",
+        "Image",
+        "Recognition"
+    ],
+    "author": "Team 4 TomorrowDevs - Refer to README for team members",
+    "license": "ISC",
+    "bugs": {
+        "url": "https://github.com/manuelalonge/team-four-webapp-frontend/issues"
+    },
+    "homepage": "https://github.com/manuelalonge/team-four-webapp-frontend#readme",
     "devDependencies": {
         "@babel/core": "^7.12.10",
         "@babel/preset-env": "^7.12.11",
         "babel-loader": "^8.2.2",
         "css-loader": "^5.0.1",
+        "html-webpack-plugin": "^4.5.1",
+        "mini-css-extract-plugin": "^1.3.4",
+        "sass": "^1.32.5",
+        "sass-loader": "^10.1.1",
+        "style-loader": "^2.0.0",
+        "webpack": "^5.17.0",
+        "webpack-cli": "^4.4.0",
+        "webpack-dev-server": "^3.11.2"
+    },
+    "dependencies": {
         "file-loader": "^6.2.0",
         "html-loader": "^1.3.2",
-        "html-webpack-plugin": "^4.5.0",
-        "mini-css-extract-plugin": "^1.3.4",
-        "sass": "^1.30.0",
-        "sass-loader": "^10.1.0",
-        "style-loader": "^2.0.0",
-        "url-loader": "^4.1.1",
-        "webpack": "^5.10.1",
-        "webpack-cli": "^4.3.1",
-        "webpack-dev-server": "^3.11.0"
+        "optimize-css-assets-webpack-plugin": "^5.0.4"
     }
 }
